refactor : MemberFixture 패키지 변경
