+++ src/test/java/online/partyrun/partyrunbattleservice/domain/member/service/MemberServiceTest.java
@@ -1,19 +1,17 @@
 package online.partyrun.partyrunbattleservice.domain.member.service;
 
-import static online.partyrun.partyrunbattleservice.fixture.MemberFixture.멤버_박성우;
-import static online.partyrun.partyrunbattleservice.fixture.MemberFixture.멤버_박현준;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import online.partyrun.partyrunbattleservice.domain.member.entity.Member;
 import online.partyrun.partyrunbattleservice.domain.member.repository.MemberRepository;
-
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.List;
+
+import static online.partyrun.partyrunbattleservice.fixture.member.MemberFixture.멤버_박성우;
+import static online.partyrun.partyrunbattleservice.fixture.member.MemberFixture.멤버_박현준;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest
 @DisplayName("MemberService")
